Add AptosGateway in Infra Category (#57)
* Add AptosGateway in Infra category

We humbly ask that you consider adding AptosGateway to the infra category; you would be greatly appreciated. We wish to continue to help the Aptos community. We intend to use the Aptos community as a hub from which it will draw energy in the future.

* Add AptosGateway in Infra

We humbly ask that you consider adding AptosGateway to the infra category; you would be greatly appreciated. We wish to continue to help the Aptos community. We intend to use the Aptos community as a hub from which it will draw energy in the future.

* Update README.md

* Add AptosGateway in Infra Category